feat(dialpad): renames users sync for more clarity
